Simple Storage of Stars On PlayerPrefs
Added simple store system for save riched stars on each level. We using two parameters for save on playerPrefs: First Parameter is storage string of current active scene. Second Parameter is storage a int value. This value intended for showing stars on LevelSelectButton. Where zero - is switch Off the star, then as a one is true, Switch on gameobject(star).